Fix some checkstyle issues
