+++ src/main/java/seedu/address/ui/PopupPanelForOrder.java
@@ -1,4 +1,4 @@
-package seedu.address.ui.popupWindow;
+package seedu.address.ui;
 
 import java.time.LocalDate;
